compile fixes

+++ src/applications/sctpapp/SCTPServer.cc
@@ -118,7 +118,7 @@ void SCTPServer::generateAndSend()
 
     cPacket* cmsg = new cPacket("CMSG");
     SCTPSimpleMessage* msg = new SCTPSimpleMessage("Server");
-    numBytes = (long uint32)par("requestLength");
+    numBytes = (uint32)par("requestLength");
     msg->setDataArraySize(numBytes);
     for (uint32 i=0; i<numBytes; i++)
     {
@@ -377,7 +377,7 @@ void SCTPServer::handleMessage(cMessage *msg)
 
                 if (echoFactor==0)
                 {
-                    if ((long uint32)par("numPacketsToReceivePerClient")>0)
+                    if ((uint32)par("numPacketsToReceivePerClient")>0)
                     {
                         j->second.rcvdPackets--;
                         SCTPSimpleMessage *smsg = check_and_cast<SCTPSimpleMessage*>(msg);
